DEKH IZHAAN TERE SE ACHA H
